:recycle: detail component move
