+++ frontend/src/views/components/detail/RecruitRequirements.tsx
